+++ valley/tests/__init__.py
@@ -1,1 +1,3 @@
 from .validators import *
+from .utils import *
+from .schema import *