+++ HISTORY.md
@@ -7,3 +7,9 @@
 ## 1.0.1
 
 - Fixed DateTimeProperty and DateTimeVariableMixin
+
+## 1.1.0
+
+- Added OrderedDeclaredVars class
+- Added more tests
+- Fixed create_error_dict flag on the BaseSchema class